Now uses miners library for MS authentication
We forked the mine-rs repository to fix some issues and add some features. The forked repository is available here: https://github.com/CCBlueX/mine-rs

It now allows us to use the MS authencation with async functions, which is a lot faster than the old method. It also fixes some issues with it.
I also improved error handling to diagnose issues.